moved internal view classes to refs
