+++ modules/core/src/view/view.ts
@@ -20,9 +20,9 @@ export function createRootView(injector: Injector, rootSelectorOrNode: string | 
 }
 
 export function createView(root: RootData, renderer: Renderer,
-  parent: ViewData, node: any): ViewData {
+  parent: ViewData, renderElement: any): ViewData {
   const view: ViewData = {
-    node,
+    renderElement,
     root,
     renderer,
     parent,
@@ -65,4 +65,8 @@ function createRootData(
     rendererFactory,
     renderer
   };
+}
+
+export function createClass(view: ViewData) {
+  // let provider = resolveReflectiveProviders([component])[0];
 }
