+++ modules/core/src/view/view_container_ref.ts
@@ -1,11 +1,8 @@
-import { ViewRef, ViewRef_ } from './view_ref';
-import { destroyView } from './view';
+import { ViewRef } from './view_ref';
 import { ElementRef } from './element_ref';
-import { ViewData } from './types';
 import { Injector } from '../di/injector';
 import { ComponentFactory } from '../component/factory';
 import { ComponentRef } from '../component/reference';
-import { createViewInjector } from './view_injector';
 
 export abstract class ViewContainerRef {
 
@@ -85,82 +82,4 @@ export abstract class ViewContainerRef {
    * If the `index` param is omitted, the last ViewRef is detached.
    */
   abstract detach(index?: number): ViewRef;
-}
-
-// tslint:disable-next-line
-class ViewContainerRef_ implements ViewContainerRef {
-
-  renderElement: any;
-  embeddedViews: ViewData[];
-
-  constructor(private _view: ViewData) { }
-
-  get anchorElement(): ElementRef { return new ElementRef(this.renderElement); }
-  get injector(): Injector { return createViewInjector(this._view); }
-  get parentInjector(): Injector {
-    return null;
-  }
-
-  clear(): void { }
-
-  get(index: number): ViewRef {
-    const view = this.embeddedViews[index];
-    if (view) {
-      const ref = new ViewRef_(view);
-      ref.attachToViewContainerRef(this);
-      return ref;
-    }
-    return null;
-  }
-
-  get length(): number { return this.embeddedViews.length; }
-
-  createEmbeddedView<C>(templateRef: any, context?: C, index?: number):
-    any { } // EmbeddedViewRef<C> { }
-
-  createComponent<C>(componentFactory: ComponentFactory<C>, index?: number, injector?: Injector,
-        rootSelectorOrNode?: any): ComponentRef<C> {
-      const contextInjector = injector || this.parentInjector;
-      const componentRef = componentFactory.create(contextInjector, rootSelectorOrNode);
-
-      return componentRef;
-    }
-
-  // insert(viewRef: ViewRef, index?: number): ViewRef { }
-
-  // move(viewRef: ViewRef, currentIndex: number): ViewRef { }
-
-  indexOf(viewRef: ViewRef): number {
-    return this.embeddedViews.indexOf((<ViewRef_>viewRef)._view);
-  }
-
-  remove(index?: number): void {
-    const view = detachEmbeddedView(this, index);
-    if (view) {
-      destroyView(view);
-    }
-  }
-
-  detach(index?: number): ViewRef {
-    const view = detachEmbeddedView(this, index);
-    return view ? new ViewRef_(view) : null;
-  }
-}
-
-export function createViewContainerRef(view: ViewData): ViewContainerRef {
-  return new ViewContainerRef_(view);
-}
-
-
-function detachEmbeddedView(container: ViewContainerRef_, viewIndex: number): ViewData {
-  const embeddedViews = container.embeddedViews;
-  if (viewIndex == null || viewIndex >= embeddedViews.length) {
-    viewIndex = embeddedViews.length - 1;
-  }
-  if (viewIndex < 0) {
-    return null;
-  }
-  const view = embeddedViews[viewIndex];
-  embeddedViews.splice(viewIndex, 1);
-  return view;
 }
